Meta関連の設定
Next.js v12で使用していたMeta情報はコメントアウトしていたので、最新バージョン(v14)に合わせたMetaデータの設定方法でtitleとdescription、canonicalを設定

metadataは'use client'を宣言しているファイルで使用できないので、ページ内の情報をん別ファイルに分割
